requirements.txt
openai
transformers
torch
Pillow
PyYAML
tqdm
pillow
python-dotenv
matplotlib
seaborn
pandas
requests
websocket-client